services/order-service/requirements.txt
fastapi
uvicorn
sqlmodel
psycopg2-binary
python-dotenv
python-jose[cryptography]
sqlalchemy